fix: address cubic-dev-ai review feedback
- Convert test file from Deno to Vitest format for CI compatibility
- Move test to correct location under src/lib/email/
- Mask sensitive invitation tokens in console logs for security
- Fix test assertions to use expect() instead of Deno asserts

Addresses all 4 issues raised in cubic-dev-ai review
