+++ public/changelog-feeds.json
@@ -1,5 +1,5 @@
 {
-  "lastGenerated": "2025-09-29T13:44:28.175Z",
+  "lastGenerated": "2025-09-30T12:09:46.173Z",
   "entriesCount": 17,
   "latestVersion": "2.1.0",
   "latestDate": "2025-08-25T00:00:00.000Z"
