+++ public/changelog-rss.xml
@@ -6,7 +6,7 @@
     <link>https://contributor.info/changelog</link>
     <atom:link href="https://contributor.info/changelog-rss.xml" rel="self" type="application/rss+xml" />
     <language>en-us</language>
-    <lastBuildDate>Mon, 29 Sep 2025 13:44:28 GMT</lastBuildDate>
+    <lastBuildDate>Tue, 30 Sep 2025 12:09:46 GMT</lastBuildDate>
     <pubDate>Mon, 25 Aug 2025 00:00:00 GMT</pubDate>
     <ttl>1440</ttl>
     <generator>contributor.info RSS Generator</generator>
